+++ p5-analysis/src/helpers/pathname-helpers.ts
@@ -1,5 +1,5 @@
 export function isHtmlPathname(pathname: string): boolean {
-  return /\.html?/i.test(pathname);
+  return /\.html?$/i.test(pathname);
 }
 
 export function isScriptPathname(pathname: string): boolean {